Moved Task-specific body implementation to body_tasks.py
